requirements.txt
ultralytics
opencv-python
numpy
pillow
gradio
pandas
matplotlib
tqdm
torch